move `impl pallet_timestamp::Config for Runtime` to separate file
