+++ runtime/devnet/src/lib.rs
@@ -866,14 +866,6 @@ impl pallet_mmr::Config for Runtime {
 	type WeightInfo = ();
 }
 
-impl pallet_timestamp::Config for Runtime {
-	/// A timestamp: milliseconds since the unix epoch.
-	type Moment = u64;
-	type OnTimestampSet = Babe;
-	type MinimumPeriod = ConstU64<{ SLOT_DURATION / 2 }>;
-	type WeightInfo = ();
-}
-
 parameter_types! {
 	pub const IndexDeposit: Balance = 1 * PAN;
 }
