+++ runtime/devnet/src/pallets/timestamp.rs
@@ -1,0 +1,11 @@
+#![allow(unused_imports)]
+
+use crate::*;
+
+impl pallet_timestamp::Config for Runtime {
+	/// A timestamp: milliseconds since the unix epoch.
+	type Moment = u64;
+	type OnTimestampSet = Babe;
+	type MinimumPeriod = ConstU64<{ SLOT_DURATION / 2 }>;
+	type WeightInfo = ();
+}
